refactor: extract render methods to RenderTrait
